Put MoteTypeInformation in GUI
This is 35 lines of code, put it inside
GUI instead of having a plugin. This
also removes the remaining special case
in the menu creation code, and an observer
of simulation.
