+++ java/org/contikios/cooja/Simulation.java
@@ -400,7 +400,8 @@ public final class Simulation extends Observable {
       }
       // Skip plugins that have been removed or merged into other classes.
       if ("org.contikios.cooja.plugins.SimControl".equals(pluginClassName) ||
-          "org.contikios.cooja.plugins.SimInformation".equals(pluginClassName)) {
+          "org.contikios.cooja.plugins.SimInformation".equals(pluginClassName) ||
+          "org.contikios.cooja.plugins.MoteTypeInformation".equals(pluginClassName)) {
         continue;
       }
       // Backwards compatibility: old visualizers were replaced.
